menstruation-predictor-app-v2/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
email-validator
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
google-genai